tc-watchdog/backend/requirements.txt
google-adk
google-generativeai
fastapi
uvicorn
python-multipart
pillow
